third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-srcset-w-onerror.html

<!DOCTYPE html>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
    }
    window.onload = function() {
        setTimeout(function() {
            if (window.testRunner)
                testRunner.notifyDone();
        }, 500);
    };
</script>
<img srcset="1x.gif 96w" onerror="console.log('Error triggerred');">
